pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "The core API using websockets to manage game state."
authors = ["Haseeb Majid <hello@haseebmajid.dev>"]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
structlog = "^21.2.0"
beanie = "^1.7.2"
fastapi = "^0.70.0"
fastapi-socketio = "^0.0.9"
uvicorn = "^0.16.0"

[tool.poetry.dev-dependencies]
tox = "^3.24.4"
black = "^21.11b1"
isort = "^5.10.1"
flake8 = "^4.0.1"
ipython = "^7.29.0"
pre-commit = "^2.15.0"
httpx = "^0.21.1"
asgi-lifespan = "^1.0.1"
pytest = "^6.2.5"
pytest-mock = "^3.6.1"
pytest-asyncio = "^0.16.0"
factory-boy = "^3.2.1"

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.black]
line-length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
